CNN/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
PyYAML>=6.0.0
medmnist>=3.0.2
